requirements.txt
fastapi==0.133.1
uvicorn==0.41.0
sqlalchemy==2.0.47
pymysql==1.1.2
pydantic==2.12.5
starlette==0.52.1
jinja2==3.1.6
python-multipart==0.0.22
itsdangerous==2.2.0
cryptography==46.0.6
PyJWT==2.10.1
httpx==0.28.1
